refactor & bugfix client missing loop

+++ include/sensor.h
@@ -1,3 +1,3 @@
 void sensorSetup();
-float getTemperature();
-float getHumidity();
+short getTemperature();
+short getHumidity();
